Added fix to packages paths
