+++ src/test/java/org/fundacionjala/pivotal/HomeTest.java
@@ -2,6 +2,7 @@ package org.fundacionjala.pivotal;
 
 import org.fundacionjala.pivotal.model.AbstractPage;
 import org.fundacionjala.pivotal.model.PageObjectFactory;
+
 import org.fundacionjala.pivotal.model.pageObject.login.HomePage;
 import org.testng.annotations.Test;
 
@@ -15,7 +16,7 @@ public class HomeTest {
     */
     @Test
     public void init() {
-        HomePage homePage = new HomePage(AbstractPage.instanceFirefoxDriver());
+        HomePage homePage = new HomePage();
         PageObjectFactory.initElements(homePage.getDriver(), homePage);
         homePage.clickOnSignIn();
         homePage.getDriver().close();
